+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Improvements
 
 - [beta] Add file scopes to exportable settings.
+- [beta] Display document icons in menus of the Back/Forward toolbar item.
 - [beta] Improve overall stability and performance.
 - [beta] Update Czech localization.